Revert comit and move QueryUtils in Kuzzle.API.QueryOptions namespace
